requirements.txt
pytest_allure_adaptor==1.7.10
xlrd==1.1.0
openpyxl==2.5.6
requests==2.19.1
Flask==1.1.2
pytest==3.8.0
kafka==1.3.5
pymysql==0.10.0
PyYAML==5.3.1
redis==3.5.3
